fix: rename folder Button to button
